chore: integrated into the main package build release flow
