+++ docs/.vitepress/theme/components/EnhanceLayout.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import SiteDebugConsole from '@docs-islands/vitepress/debug-console';
 import { Analytics } from '@vercel/analytics/vue';
 import DefaultTheme from 'vitepress/theme';
 </script>
@@ -7,5 +6,4 @@ import DefaultTheme from 'vitepress/theme';
 <template>
   <DefaultTheme.Layout />
   <Analytics />
-  <SiteDebugConsole />
 </template>
