--- docs/debug-console.d.ts
@@ -1,7 +1,0 @@
-import type { DefineComponent } from 'vue';
-
-declare module '@docs-islands/vitepress/debug-console' {
-  const SiteDebugConsole: DefineComponent;
-
-  export default SiteDebugConsole;
-}
