+++ packages/vitepress/docs/.vitepress/theme/components/EnhanceLayout.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import SiteDebugConsole from '@docs-islands/vitepress/debug-console';
+import SiteDebugConsole from '@docs-islands/vitepress/debug-console/client';
+import '@docs-islands/vitepress/debug-console/client/style.css';
 import { Analytics } from '@vercel/analytics/vue';
 import DefaultTheme from 'vitepress/theme';
 </script>
