+++ packages/vitepress/tsconfig.json
@@ -18,7 +18,6 @@
     "tests/",
     "scripts/",
     "theme/**/*.ts",
-    "theme/**/*.d.ts",
     "*.ts",
   ],
 }
