+++ packages/vitepress/types/client.d.ts
@@ -1,8 +1,3 @@
-declare module '*.css' {
-  const css: string;
-  export default css;
-}
-
 declare module '*.vue' {
   import type { DefineComponent } from 'vue';
