+++ packages/vitepress/types/debug-console.d.ts
@@ -2,4 +2,5 @@ import type { DefineComponent } from 'vue';
 
 declare const SiteDebugConsole: DefineComponent;
 
+// Vue module type generation failed, details can be seen in the issue: https://github.com/sxzz/rolldown-plugin-dts/issues/188
 export default SiteDebugConsole;
