+++ packages/vitepress/types/theme.css.d.ts
@@ -1,0 +1,1 @@
+declare module '@docs-islands/vitepress/debug-console/client/style.css' {}
