+++ pnpm-workspace.yaml
@@ -159,9 +159,10 @@ catalogs:
     '@babel/traverse': '^7.29.0'
     '@babel/types': '^7.29.0'
     'rolldown': '1.0.0-rc.4'
-    'rolldown-plugin-dts': '^0.22.1'
+    'rolldown-plugin-dts': '^0.23.1'
     'rollup-plugin-license': '^3.6.0'
     '@vitejs/plugin-react-swc': '^4.2.3'
+    '@vitejs/plugin-vue': '^6.0.5'
     'concurrently': '^9.2.1'
     'is-in-ci': '^2.0.0'
     'shiki': ^2.5.0
