+++ utils/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "isolatedDeclarations": true,
     "declaration": true,
-    "lib": ["ESNext", "DOM"],
+    "lib": ["ESNext", "DOM", "DOM.Iterable"],
     "types": ["node"],
   },
   "include": ["./", "./logger.d.ts"],
